+++ src/td/ipxmgr.h
@@ -163,7 +163,7 @@ it's stored in protected-mode memory.
 ---------------------------------------------------------------------------*/
 extern "C" {
 void* Get_RM_IPX_Address();
-long Get_RM_IPX_Size();
+int32_t Get_RM_IPX_Size();
 }
 
 /*
@@ -358,7 +358,7 @@ class IPXManagerClass : public ConnManClass {
   This is a real-mode pointer to the address of the real-mode assembly
   entry point.
   .....................................................................*/
-  long Handler = 0;
+  int32_t Handler = 0;
 
   /*.....................................................................
   Event Control Block for listening; contained within the real-mode
